--- alembic/versions/1291b7b7903c_mensagem.py
@@ -1,28 +1,0 @@
-"""<mensagem>
-
-Revision ID: 1291b7b7903c
-Revises: ac892b9eb2e4
-Create Date: 2023-08-17 00:06:34.221628
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1291b7b7903c'
-down_revision = 'ac892b9eb2e4'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
