--- alembic/versions/a3dbe3fd3963_alternado_o_tamanho_da_coluna_senha.py
@@ -1,27 +1,0 @@
-"""alternado o tamanho da coluna senha
-
-Revision ID: a3dbe3fd3963
-Revises: c7aa6a0ac7d8
-Create Date: 2023-08-15 08:42:18.697967
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a3dbe3fd3963'
-down_revision = 'c7aa6a0ac7d8'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('usuario', 'senha', type_=sa.String(length=200))
-    op.alter_column('loja', 'senha', type_=sa.String(length=200))
-
-
-def downgrade() -> None:
-    op.alter_column('usuario', 'senha', type_=sa.String(length=50))
-    op.alter_column('loja', 'senha', type_=sa.String(length=200))
