+++ alembic/versions/bbf748d9b646_versão_final_do_banco.py
@@ -1,0 +1,110 @@
+"""versão final do banco
+
+Revision ID: bbf748d9b646
+Revises: 
+Create Date: 2023-08-17 15:16:24.099618
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'bbf748d9b646'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('loja',
+    sa.Column('cnpj', sa.String(length=14), nullable=False),
+    sa.Column('email', sa.String(length=50), nullable=True),
+    sa.Column('senha', sa.String(length=200), nullable=True),
+    sa.Column('nome', sa.String(length=100), nullable=True),
+    sa.PrimaryKeyConstraint('cnpj')
+    )
+    op.create_table('status',
+    sa.Column('id_status', sa.Integer(), nullable=False),
+    sa.Column('descricao', sa.String(length=100), nullable=True),
+    sa.PrimaryKeyConstraint('id_status')
+    )
+    op.create_table('usuario',
+    sa.Column('cpf', sa.String(length=11), nullable=False),
+    sa.Column('nome', sa.String(length=100), nullable=True),
+    sa.Column('email', sa.String(length=50), nullable=True),
+    sa.Column('senha', sa.String(length=200), nullable=True),
+    sa.Column('admin', sa.Boolean(), nullable=True),
+    sa.PrimaryKeyConstraint('cpf')
+    )
+    op.create_table('endereco',
+    sa.Column('cep', sa.String(length=8), nullable=False),
+    sa.Column('rua', sa.String(length=50), nullable=True),
+    sa.Column('numero', sa.Integer(), nullable=False),
+    sa.Column('complemento', sa.String(length=50), nullable=False),
+    sa.Column('cidade', sa.String(length=50), nullable=True),
+    sa.Column('estado', sa.String(length=50), nullable=True),
+    sa.Column('cpf_usuario', sa.String(length=11), nullable=True),
+    sa.ForeignKeyConstraint(['cpf_usuario'], ['usuario.cpf'], ),
+    sa.PrimaryKeyConstraint('cep', 'numero', 'complemento')
+    )
+    op.create_table('pedido',
+    sa.Column('id_pedido', sa.Integer(), nullable=False),
+    sa.Column('cpf_usuario', sa.String(length=11), nullable=True),
+    sa.Column('preco_total', sa.Float(), nullable=True),
+    sa.Column('pedido_status', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['cpf_usuario'], ['usuario.cpf'], ),
+    sa.ForeignKeyConstraint(['pedido_status'], ['status.id_status'], ),
+    sa.PrimaryKeyConstraint('id_pedido')
+    )
+    op.create_table('produto',
+    sa.Column('id_produto', sa.Integer(), nullable=False),
+    sa.Column('cnpj_loja', sa.String(length=14), nullable=True),
+    sa.Column('categoria_prod', sa.String(length=50), nullable=True),
+    sa.Column('nome_produto', sa.String(length=50), nullable=True),
+    sa.Column('marca_produto', sa.String(length=50), nullable=True),
+    sa.Column('preco', sa.String(length=50), nullable=True),
+    sa.Column('especificacoes', sa.String(length=100), nullable=True),
+    sa.ForeignKeyConstraint(['cnpj_loja'], ['loja.cnpj'], ),
+    sa.PrimaryKeyConstraint('id_produto')
+    )
+    op.create_table('telefone',
+    sa.Column('cpf_usuario', sa.String(length=11), nullable=True),
+    sa.Column('numero', sa.String(), nullable=False),
+    sa.ForeignKeyConstraint(['cpf_usuario'], ['usuario.cpf'], ),
+    sa.PrimaryKeyConstraint('numero')
+    )
+    op.create_table('estoque',
+    sa.Column('id_estoque', sa.Integer(), nullable=False),
+    sa.Column('id_produto', sa.Integer(), nullable=True),
+    sa.Column('id_loja', sa.String(length=14), nullable=True),
+    sa.Column('quantidade', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['id_loja'], ['loja.cnpj'], ),
+    sa.ForeignKeyConstraint(['id_produto'], ['produto.id_produto'], ),
+    sa.PrimaryKeyConstraint('id_estoque')
+    )
+    op.create_table('item',
+    sa.Column('id_item', sa.Integer(), nullable=False),
+    sa.Column('id_produto', sa.Integer(), nullable=True),
+    sa.Column('id_pedido', sa.Integer(), nullable=True),
+    sa.Column('quantidade', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['id_pedido'], ['pedido.id_pedido'], ),
+    sa.ForeignKeyConstraint(['id_produto'], ['produto.id_produto'], ),
+    sa.PrimaryKeyConstraint('id_item')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('item')
+    op.drop_table('estoque')
+    op.drop_table('telefone')
+    op.drop_table('produto')
+    op.drop_table('pedido')
+    op.drop_table('endereco')
+    op.drop_table('usuario')
+    op.drop_table('status')
+    op.drop_table('loja')
+    # ### end Alembic commands ###
