--- alembic/versions/c7aa6a0ac7d8_banco_com_estoque_apontando_loja.py
@@ -1,30 +1,0 @@
-"""banco com estoque apontando loja
-
-Revision ID: c7aa6a0ac7d8
-Revises: 69e120cd2870
-Create Date: 2023-08-11 20:59:25.328030
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'c7aa6a0ac7d8'
-down_revision = '69e120cd2870'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('estoque', sa.Column('id_loja', sa.String(length=14), nullable=True))
-    op.create_foreign_key(None, 'estoque', 'loja', ['id_loja'], ['cnpj'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'estoque', type_='foreignkey')
-    op.drop_column('estoque', 'id_loja')
-    # ### end Alembic commands ###
